requirements.txt
crewai
langchain
langchain-openai
langchain-chroma
pydantic
pyyaml
python-dotenv
requests  # For WeatherAPI and Booking API calls
thefuzz[speedup]  # For fuzzy string matching (destination detection)
dateparser  # For intelligent date parsing
pandas  # For data manipulation